inference

+++ diffusion/gaussian_diffusion.py
@@ -418,7 +418,7 @@ class GaussianDiffusion:
         if cond_fn is not None:
             out["mean"] = self.condition_mean(cond_fn, out, x, t, model_kwargs=model_kwargs)
         sample = out["mean"] + nonzero_mask * th.exp(0.5 * out["log_variance"]) * noise
-        return {"sample": sample, "pred_xstart": out["pred_xstart"]}
+        return {"sample": sample.to(dtype=x.dtype), "pred_xstart": out["pred_xstart"]}
 
     def p_sample_loop(
         self,
